backend/requirements.txt
fastapi>=0.115.7
uvicorn[standard]>=0.23.1
python-multipart>=0.0.18
pydantic>=2.8.0
pydantic-settings>=2.2.1
motor>=3.4.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.10.1
email-validator>=1.3.0
